storage: handle authorization in the client
Add option to set auth key.

Ensure authorized before get/put. This does nothing if the authorized
flag is set to true (already authorized) or the auth key is empty
(indicating no need for authorization). Otherwise it sends an auth
message and either receives an auth message in response, in which
case the authorized flag is set to true, or an error is returned
and the connection is not authorized.

Handle "go away" response for get/put, clearing the authorized flag.
This can happen on connection loss/reconnect of the underlying client.